requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.177.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via gunicorn
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   google-generativeai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pyparsing==3.2.3
    # via httplib2
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
requests==2.32.4
    # via
    #   -r requirements.in
    #   google-api-core
rsa==4.9.1
    # via google-auth
ruff==0.12.5
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp
